inline some stuff
