+++ crates/mun_codegen/src/code_gen.rs
@@ -1,4 +1,4 @@
-use crate::type_info::TypeManager;
+use crate::ir::ty::TypeManager;
 use crate::ir::file::FileIR;
 use crate::ir::file_group::FileGroupIR;
 use crate::code_gen::linker::LinkerError;
