+++ crates/mun_codegen/src/ir/ty.rs
@@ -1,3 +1,6 @@
+use std::collections::HashMap;
+use mun_target::abi::TargetDataLayout;
+use inkwell::context::Context;
 use super::try_convert_any_to_basic;
 use crate::{
     type_info::{TypeInfo, TypeSize},
@@ -16,11 +19,12 @@ use inkwell::{
 #[rustfmt::skip]
 pub(crate) fn ir_query(db: &impl IrDatabase, ty: Ty, params: CodeGenParams) -> AnyTypeEnum {
     let context = db.context();
+    let layout = db.target_data_layout();
     match ty {
         Ty::Empty => AnyTypeEnum::StructType(context.struct_type(&[], false)),
         Ty::Apply(ApplicationTy { ctor, .. }) => match ctor {
-            TypeCtor::Float(fty) => float_ty_query(db, fty).into(),
-            TypeCtor::Int(ity) => int_ty_query(db, ity).into(),
+            TypeCtor::Float(fty) => float_ty_query(context.as_ref(), &layout, fty).into(),
+            TypeCtor::Int(ity) => int_ty_query(context.as_ref(), &layout, ity).into(),
             TypeCtor::Bool => AnyTypeEnum::IntType(context.bool_type()),
 
             TypeCtor::FnDef(def @ CallableDef::Function(_)) => {
@@ -57,19 +61,76 @@ pub(crate) fn ir_query(db: &impl IrDatabase, ty: Ty, params: CodeGenParams) -> A
     }
 }
 
+#[derive(Debug)]
+pub struct TypeManager {
+    infos: HashMap<hir::Ty, TypeInfo>,
+}
+
+impl TypeManager {
+    pub fn new() -> TypeManager {
+        TypeManager {
+            infos: HashMap::new(),
+        }
+    }
+
+    pub fn type_info<D: IrDatabase>(&mut self, db: &D, ty: hir::Ty) -> TypeInfo {
+        if let Some(info) = self.infos.get(&ty) {
+            return info.clone();
+        }
+
+        let context = db.context();
+        let target = db.target_data();
+        let layout = db.target_data_layout();
+ 
+        let res = match &ty {
+            Ty::Apply(ctor) => match ctor.ctor {
+                TypeCtor::Float(ty) => {
+                    let ir_ty = float_ty_query(&context, &layout, ty);
+                    let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
+                    TypeInfo::new_fundamental(
+                        format!("core::{}", ty.resolve(&layout)),
+                        type_size,
+                    )
+                }
+                TypeCtor::Int(ty) => {
+                    let ir_ty = int_ty_query(&context, &layout, ty);
+                    let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
+                    TypeInfo::new_fundamental(
+                        format!("core::{}", ty.resolve(&layout)),
+                        type_size,
+                    )
+                }
+                TypeCtor::Bool => {
+                    let ir_ty = context.bool_type();
+                    let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
+                    TypeInfo::new_fundamental("core::bool", type_size)
+                }
+                TypeCtor::Struct(s) => {
+                    let ir_ty = db.struct_ty(s);
+                    let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
+                    return TypeInfo::new_struct(db, s, type_size)
+                }
+                _ => unreachable!("{:?} unhandled", ctor),
+            },
+            _ => unreachable!("{:?} unhandled", ty),
+        };
+
+        assert!(self.infos.insert(ty, res.clone()).is_none());
+        res
+    }
+}
+
 /// Returns the LLVM IR type of the specified float type
-fn float_ty_query(db: &impl IrDatabase, fty: FloatTy) -> FloatType {
-    let context = db.context();
-    match fty.bitness.resolve(&db.target_data_layout()) {
+fn float_ty_query(context: &Context, layout: &TargetDataLayout, fty: FloatTy) -> FloatType {
+    match fty.bitness.resolve(layout) {
         FloatBitness::X64 => context.f64_type(),
         FloatBitness::X32 => context.f32_type(),
     }
 }
 
 /// Returns the LLVM IR type of the specified int type
-fn int_ty_query(db: &impl IrDatabase, ity: IntTy) -> IntType {
-    let context = db.context();
-    match ity.bitness.resolve(&db.target_data_layout()) {
+fn int_ty_query(context: &Context, layout: &TargetDataLayout, ity: IntTy) -> IntType {
+    match ity.bitness.resolve(layout) {
         IntBitness::X128 => context.i128_type(),
         IntBitness::X64 => context.i64_type(),
         IntBitness::X32 => context.i32_type(),
@@ -93,41 +154,4 @@ pub fn struct_ty_query(db: &impl IrDatabase, s: hir::Struct) -> StructType {
     }
 
     db.context().opaque_struct_type(&name)
-}
-
-/// Constructs the `TypeInfo` for the specified HIR type
-pub fn type_info_query(db: &impl IrDatabase, ty: Ty) -> TypeInfo {
-    let target = db.target_data();
-    match ty {
-        Ty::Apply(ctor) => match ctor.ctor {
-            TypeCtor::Float(ty) => {
-                let ir_ty = float_ty_query(db, ty);
-                let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
-                TypeInfo::new_fundamental(
-                    format!("core::{}", ty.resolve(&db.target_data_layout())),
-                    type_size,
-                )
-            }
-            TypeCtor::Int(ty) => {
-                let ir_ty = int_ty_query(db, ty);
-                let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
-                TypeInfo::new_fundamental(
-                    format!("core::{}", ty.resolve(&db.target_data_layout())),
-                    type_size,
-                )
-            }
-            TypeCtor::Bool => {
-                let ir_ty = db.context().bool_type();
-                let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
-                TypeInfo::new_fundamental("core::bool", type_size)
-            }
-            TypeCtor::Struct(s) => {
-                let ir_ty = db.struct_ty(s);
-                let type_size = TypeSize::from_ir_type(&ir_ty, target.as_ref());
-                TypeInfo::new_struct(db, s, type_size)
-            }
-            _ => unreachable!("{:?} unhandled", ctor),
-        },
-        _ => unreachable!("{:?} unhandled", ty),
-    }
 }
